# internal/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PlaylistLabs User Management API",
    "description": "REST API for PlaylistLabs User Management open-source panel.\nAllows third-party systems, and automation tools to manage **Playlists & Sync**, **Managed Users**, **Team Members / Collaborators**, and **API Tokens**.\n\n### Authentication\nAuthentication is performed strictly using an internal API Token (`plt_live_...`) in the canonical **`X-API-Token`** header:\n- Header: `X-API-Token: <your_token>`\n\n### Optional Token Password\nIf the API Token was configured with an optional password, include it in the **`X-Token-Password`** header:\n- Header: `X-Token-Password: <your_password>`\n\n### Anti-Brute-Force & IP Banning\nInvalid tokens or repeated incorrect token passwords trigger automatic IP rate tracking.\nExceeding the maximum allowed failed attempts within the security window causes an **immediate IP ban** (`403 Forbidden`). Once banned, subsequent requests from the IP are rejected with no further logging.\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server base API path"
    }
  ],
  "security": [
    {
      "ApiTokenAuth": []
    },
    {
      "ApiTokenAuth": [],
      "TokenPasswordAuth": []
    }
  ],
  "paths": {
    "/playlists": {
      "get": {
        "summary": "List accessible playlists",
        "description": "Returns all playlists accessible to the caller. Automatically scoped according to API token playlist restrictions (`allowed_playlist_ids`) and collaborator assignments.",
        "operationId": "getPlaylists",
        "tags": [
          "Playlists"
        ],
        "responses": {
          "200": {
            "description": "List of accessible playlists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Playlist"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/playlists/{listId}/refresh": {
      "post": {
        "summary": "Refresh playlist from PlaylistLabs",
        "description": "Triggers an asynchronous background synchronization job with PlaylistLabs for the specified playlist.\nFetches updated channels, VODs, series, and categories from PlaylistLabs and reconciles them in MariaDB.\nSubject to token playlist scope checks (`allowed_playlist_ids`).\n",
        "operationId": "refreshPlaylist",
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "description": "Playlist ID (64-bit integer)",
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1001
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Whether to force a full re-sync of all streams and categories (default: false)",
            "schema": {
              "type": "boolean",
              "default": false,
              "example": false
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "Optional JSON body specifying sync options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Synchronization job accepted and started in background",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSyncResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Another synchronization job is already running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictResponse"
                }
              }
            }
          },
          "503": {
            "description": "Synchronization engine is not available on this server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{listId}/sync": {
      "post": {
        "summary": "Synchronize playlist with PlaylistLabs (alias)",
        "description": "Identical alias to `POST /playlists/{listId}/refresh`.",
        "operationId": "syncPlaylist",
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "description": "Playlist ID (64-bit integer)",
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1001
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Whether to force a full re-sync of all streams and categories (default: false)",
            "schema": {
              "type": "boolean",
              "default": false,
              "example": false
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Synchronization job accepted and started in background",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSyncResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Another synchronization job is already running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictResponse"
                }
              }
            }
          },
          "503": {
            "description": "Synchronization engine is not available on this server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/sync/status": {
      "get": {
        "summary": "Get sync engine status",
        "description": "Returns the current execution status of the background synchronization engine, including the active job ID, current step, and running status.",
        "operationId": "getSyncStatus",
        "tags": [
          "Playlists"
        ],
        "responses": {
          "200": {
            "description": "Current status of the sync engine",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/playlists/{listId}/import-from-editor": {
      "post": {
        "summary": "Import users and welcome info from PlaylistLabs",
        "description": "Directly imports customer accounts (managed users) and welcome information configured on PlaylistLabs for this playlist.",
        "operationId": "importFromEditor",
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "description": "Playlist ID (64-bit integer)",
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1001
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportFromEditorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFromEditorResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Failed connecting to PlaylistLabs or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/sync": {
      "post": {
        "summary": "Synchronize all playlists with PlaylistLabs",
        "description": "Triggers an asynchronous background synchronization job across all playlists accessible to the caller.\n",
        "operationId": "syncAllPlaylists",
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Whether to force a full re-sync of all streams and categories (default: false)",
            "schema": {
              "type": "boolean",
              "default": false,
              "example": false
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "Optional JSON body specifying sync options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Synchronization job accepted and started in background",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSyncResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Another synchronization job is already running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictResponse"
                }
              }
            }
          },
          "503": {
            "description": "Synchronization engine is not available on this server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/refresh": {
      "post": {
        "summary": "Refresh all playlists from PlaylistLabs (alias)",
        "description": "Identical alias to `POST /playlists/sync`.",
        "operationId": "refreshAllPlaylists",
        "tags": [
          "Playlists"
        ],
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Whether to force a full re-sync of all streams and categories (default: false)",
            "schema": {
              "type": "boolean",
              "default": false,
              "example": false
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "Optional JSON body specifying sync options",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Synchronization job accepted and started in background",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSyncResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "description": "Another synchronization job is already running",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictResponse"
                }
              }
            }
          },
          "503": {
            "description": "Synchronization engine is not available on this server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "PlaylistLabs internal API Token (format: plt_live_...)"
      },
      "TokenPasswordAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-Password",
        "description": "Optional password configured during API token generation (leave empty if no password was set)"
      }
    },
    "schemas": {
...
      "SyncRequest": {
        "type": "object",
        "properties": {
          "force": {
            "type": "boolean",
            "description": "Force complete re-sync of all streams and categories",
            "default": true,
            "example": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters or body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed, missing token, expired token, or invalid token password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied, insufficient permissions, API tokens disabled for this account, or IP address blocked",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}